Improve shell scripts to start quake3
The ioquake3 argv parser is dumb and treats spaces and NULLs the same.
So adding something without + to the argv, would have it appended to the
previous argument (+set com_basegame baseq3), causing it to look in e.g.
'~/.q3a/baseq3 arg1'.

(The argv parsing is likely a remnant of old windows/msdos behaviour.)